\name{mixSPE-package}
\alias{mixSPE-package}
\alias{mixSPE}
\docType{package}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Mixtures of skew power exponential or power exponential distributions.
}
\description{
An implementation of skewed and elliptical mixture distributions for use in model-based clustering.
}
%- maybe also 'usage' for other objects documented here.
\details{
\tabular{ll}{
Package: \tab mixSPE\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2019-06-18\cr
License: \tab GPL (>= 2)\cr
}
}
\keyword{ package }
