\name{print.pemix}
\alias{print.pemix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print a summary of the model fit.
}
\description{
Print a summary of the model fit including the number of components and the scale structure selected by the BIC and the ICL.}

\usage{
\method{print}{pemix}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of class "pemix".}
  \item{\dots}{
Ignore this
}
}

\author{
Utkarsh J. Dang, Ryan P. Browne, and Paul D. McNicholas
}
