\name{print.spemix}
\alias{print.spemix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print a summary of the model fit.
}
\description{
Print a summary of the model fit including the number of components and the scale structure selected by the BIC and the ICL.}
\usage{
\method{print}{spemix}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of class "spemix".}
  \item{\dots}{
Ignore this
}
}
\author{
Utkarsh J. Dang, Michael P. B. Gallaugher, Ryan P. Browne, and Paul D. McNicholas
}


